BACK-END API CREATION USING PYTON-FLASK
The above three text files are everything about creation of API's that are given as a task. Among them, two of them are Main files such as First API & Report generation which are the API's or Apps that consists the code. And the third one is something which indicates the way to be followed for loading CSV files in to a database.